pyproject.toml
[tool.poetry]
name = "editoggia-v1"
version = "1.0.0"
description = "Simple fanfiction site"
authors = ["Louise <louise.aur@mailbox.org>"]
license = "MPL"

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.1.1"
Flask-SQLAlchemy = "^2.5.1"
Flask-Migrate = "^3.1.0"
Flask-Admin = "^1.6.0"
Flask-BabelEx = "^0.9.4"
Flask-Login = "^0.6.0"
Flask-WTF = "^1.0.1"
libsass = "^0.21.0"
WTForms = {extras = ["email"], version = "^3.0.1"}
email-validator = "^1.1.3"
Flask-Bcrypt = "^1.0.1"
requests = "^2.27.1"
arrow = "^1.2.2"
bleach = "^5.0.0"
bs4 = "^0.0.1"
psycopg2 = "^2.9.3"
Flask-Assets = "^2.0"
gunicorn = {version = "^20.1.0", optional = true}

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
coverage = "^6.3.2"
isort = "^5.10.1"
black = "^22.3.0"
Faker = "^13.4.0"
types-psycopg2 = "^2.9.13"

[tool.poetry.extras]
production = ["gunicorn"]

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
